// src/app/api/swagger/models/api-application-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiApplicationDocumentDto } from '../models/api-application-document-dto';
import { ApiApplicationSourceEnum } from '../models/api-application-source-enum';
import { ApiApplicationStateEnum } from '../models/api-application-state-enum';
import { ApiBankingDetailsDto } from '../models/api-banking-details-dto';
import { ApiContactInfoDto } from '../models/api-contact-info-dto';
import { ApiOrganizationCountryAltEnum } from '../models/api-organization-country';
import { ApiProductDto } from '../models/api-product-dto';
import { ApiTradeBusinessFlowEnum } from '../models/api-trade-business-flow-enum';
import { ApiTradeStateEnum } from '../models/api-trade-state-enum';
import { ApiVatRateStateEnum } from '../models/api-vat-rate-state-enum';
export interface ApiApplicationDto {
  AccountNumber?: string | null;
  AgreeWithCustomerConditions?: boolean;
  ApplicationDocuments?: Array<ApiApplicationDocumentDto> | null;
  ApplicationSource?: ApiApplicationSourceEnum | null;
  ApplicationState?: ApiApplicationStateEnum;
  ApplicationStateDescription?: string | null;
  ApplicationStateId?: number | null;
  BankingDetails?: ApiBankingDetailsDto | null;
  CanDelete?: boolean;
  CanEdit?: boolean;
  CanView?: boolean;
  CanWithdraw?: boolean;
  ConditionsOfPayment?: string | null;
  ContactInfo?: ApiContactInfoDto | null;
  CreatedByCustomer?: boolean;
  DefineCustomPriceForEachProduct?: boolean;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  EmailForReceipt?: string | null;
  EnableKtruSpecifications?: boolean;
  EnablePositionCharacteristics?: boolean;
  Host?: string | null;
  Id?: number;
  IncomplianceWithTheRequirementsReason?: string | null;
  IsApplicationFromAnotherTenant?: boolean;
  IsHidePrice?: boolean;
  IsLotPriceWithVat?: boolean;
  IsOrderApplication?: boolean;
  IsPriceCorridorExclusion?: boolean | null;
  IsPriceCorridorMatching?: boolean | null;
  IsPublicOffer?: boolean;
  IsRequirementsConfirmationCheck?: boolean;
  IsUnitBidding?: boolean;
  MatchCustomerRequirements44Documents?: Array<ApiApplicationDocumentDto> | null;
  NeedEditOnMarket?: boolean;
  NeverExpired?: boolean;
  OfferExpiryDate?: string | null;
  PayCommissionReportGuid?: string;
  Price?: number | null;
  Products?: Array<ApiProductDto> | null;
  PublicationDate?: string | null;
  RowVersion?: number;
  ShipmentPoint?: string | null;
  SmallBusinessRequirementDocuments?: Array<ApiApplicationDocumentDto> | null;
  SubmitterId?: string | null;
  SubmitterName?: string | null;
  SubmitterOrganizationCountry?: ApiOrganizationCountryAltEnum;
  TenantSource?: number | null;
  TradeBusinessFlow?: ApiTradeBusinessFlowEnum;
  TradeId?: number;
  TradeLotId?: number;
  TradeOrganizerId?: number | null;
  TradeState?: ApiTradeStateEnum;
  UniversalRequirementDocuments?: Array<ApiApplicationDocumentDto> | null;
  VatRate?: number | null;
  VatRateDescription?: string | null;
  VatRateState?: ApiVatRateStateEnum;
  WasRejected?: boolean;
}