// src/app/api/ui/models/application.interface.ts
import { ApplicationDocument } from './application-document.interface';
import { ApplicationSourceEnum } from '../enums/application-source.enum';
import { ApplicationStateEnum } from '../enums/application-state.enum';
import { BankingDetails } from './banking-details.interface';
import { ContactInfo } from './contact-info.interface';
import { OrganizationCountryAltEnum } from '../enums/organization-country-alt.enum';
import { Product } from './product.interface';
import { TradeBusinessFlowEnum } from '../enums/trade-business-flow.enum';
import { TradeStateEnum } from '../enums/trade-state.enum';
import { VatRateStateEnum } from '../enums/vat-rate-state.enum';

export interface Application {
  accountNumber: string;
  agreeWithCustomerConditions: boolean;
  applicationDocuments: ApplicationDocument[];
  applicationSource: ApplicationSourceEnum;
  applicationState: ApplicationStateEnum;
  applicationStateDescription: string;
  applicationStateId: number;
  bankingDetails: BankingDetails;
  canDelete: boolean;
  canEdit: boolean;
  canView: boolean;
  canWithdraw: boolean;
  conditionsOfPayment: string;
  contactInfo: ContactInfo;
  createdByCustomer: boolean;
  defineCustomPriceForEachProduct: boolean;
  deliveryPlace: string;
  deliveryTerms: string;
  emailForReceipt: string;
  enableKtruSpecifications: boolean;
  enablePositionCharacteristics: boolean;
  host: string;
  id: number;
  incomplianceWithTheRequirementsReason: string;
  isApplicationFromAnotherTenant: boolean;
  isHidePrice: boolean;
  isLotPriceWithVat: boolean;
  isOrderApplication: boolean;
  isPriceCorridorExclusion: boolean;
  isPriceCorridorMatching: boolean;
  isPublicOffer: boolean;
  isRequirementsConfirmationCheck: boolean;
  isUnitBidding: boolean;
  matchCustomerRequirements44Documents: ApplicationDocument[];
  needEditOnMarket: boolean;
  neverExpired: boolean;
  offerExpiryDate: string;
  payCommissionReportGuid: string;
  price: number;
  products: Product[];
  publicationDate: string;
  rowVersion: number;
  shipmentPoint: string;
  smallBusinessRequirementDocuments: ApplicationDocument[];
  submitterId: string;
  submitterName: string;
  submitterOrganizationCountry: OrganizationCountryAltEnum;
  tenantSource: number;
  tradeBusinessFlow: TradeBusinessFlowEnum;
  tradeId: number;
  tradeLotId: number;
  tradeOrganizerId: number;
  tradeState: TradeStateEnum;
  universalRequirementDocuments: ApplicationDocument[];
  vatRate: number;
  vatRateDescription: string;
  vatRateState: VatRateStateEnum;
  wasRejected: boolean;
}